backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
yfinance==0.2.28
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
requests>=2.31.0
gunicorn>=21.2.0
alpaca-trade-api==3.2.0
